requirements.txt
gradio>=4.19.2
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.1
openai>=1.3.0
aiohttp>=3.9.1
langchain-groq>=0.0.1